requirements.txt
drf-yasg
Django~=6.0
djangorestframework~=3.16.1
PyJWT~=2.10.1